+++ lib/framework/HealthCheckService.h
@@ -7,7 +7,7 @@
  *   Provides a /rest/health endpoint with callback-based extensibility
  *   for the ESP32 SvelteKit framework.
  *
- *   Copyright (C) 2025 hmbacher
+ *   Copyright (C) 2026 hmbacher
  *
  *   All Rights Reserved. This software may be modified and distributed under
  *   the terms of the LGPL v3 license. See the LICENSE file for details.