+++ scripts/generate_enums.py
@@ -1,7 +1,7 @@
 # Generate TypeScript enums from C++ header files.
 # This script parses C++ enum definitions and creates corresponding TypeScript enums.
 # Can be used both as standalone script and as PlatformIO pre-action.
-# Copyright (C) 2025 hmbacher
+# Copyright (C) 2026 hmbacher
 #
 # All Rights Reserved. This software may be modified and distributed under
 # the terms of the LGPL v3 license. See the LICENSE file for details.